+++ NEWS.md
@@ -1,5 +1,11 @@
 # `chrisr` 0.99
 
+## Changes in 0.99.3
+
+- Add functionality to extract CTFF-compliant data from a
+  `SummarizedExperiment`: methods `chris_data`, `chris_labels` and
+  `chris_groups`.
+
 ## Changes in 0.99.2
 
 - Add documentation and fix small issues.